Be friendlier to users who mess up the IS assertion
Previously, if one passed a custom message that evaluated to NIL, or a
wrong number of arguments, a cryptic DESCRIBE-OBJECT error internal to
FIASCO would be issued.  Now the failed assertion contains a helpful
message pointing to the user's oversight.

* src/infrastructure.lisp (failed-assertion): Delete slots already
found in CONDITION and accessible by
SIMPLE-CONDITION-FORMAT-{CONTROL|ARGUMENTS}
(describe-object): Use SIMPLE-CONDITION-FORMAT-CONTROL and
SIMPLE-CONDITION-FORMAT-ARGUMENTS.