Fix skipping tests
Although skipping tests in SBCL behaves as expected, in ACL a skipped
test still executes the test and all its assertions.

The real problem lies in the fact that ACL also provide a CONTINUE
restart when WARN is used.

Demote the TEST-SKIPPED condition from a warning (there is no reason
why it should be one) and correctly handle the skipping of a test by
explicitly invoking a SKIP-TEST restart.